# Speeding up processing the AssetCollection